requirements.txt
# Echoes Data Pipeline - Phase 1 + Phase 2 Dependencies
# Python 3.11+ required

# Reddit API
praw>=7.7.0

# LLM SDKs
google-generativeai>=0.8.0
openai>=1.50.0

# Vector Database
qdrant-client>=1.12.0

# Data Modeling & Validation
pydantic>=2.0
pydantic-settings>=2.0

# Environment & Config
python-dotenv>=1.0

# SQLite utilities
sqlite-utils>=3.35

# CLI & Logging
rich>=13.0

# Retry logic
tenacity>=8.0

# Async support
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# --- Phase 2: Personality Engine ---

# PostgreSQL (async)
asyncpg>=0.29.0

# PostgreSQL (sync - for migrations & scripts)
psycopg2-binary>=2.9.9

# Math (cosine similarity)
numpy>=1.26.0

# --- Phase 3: RAG Core ---
# (Uses existing deps: qdrant-client, google-generativeai, asyncpg, numpy, pydantic)

# Streaming HTTP for LLM responses
httpx>=0.27.0
